Hide integration tests behind the integration build tag.
