+++ pkg/test/case_test.go
@@ -220,15 +220,3 @@ func TestCollectTestStepFiles(t *testing.T) {
 		})
 	}
 }
-
-func TestRunWithKudo(t *testing.T) {
-	harness := Harness{
-		T:                 t,
-		CRDDir:            "../../config/crds/",
-		ManifestsDir:      "../../config/samples/test-framework/",
-		TestDirs:          []string{"./test_data/"},
-		StartControlPlane: true,
-	}
-
-	harness.Run()
-}
