+++ pkg/test/utils/kubernetes_test.go
@@ -1,16 +1,11 @@
 package utils
 
 import (
-	"context"
-	"fmt"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/envtest"
 )
 
 func TestNamespaced(t *testing.T) {
@@ -49,51 +44,5 @@ func TestNamespaced(t *testing.T) {
 			assert.Equal(t, namespace, actualNamespace)
 			assert.Equal(t, namespace, m.GetNamespace())
 		})
-	}
-}
-
-func TestCreateOrUpdate(t *testing.T) {
-	env := &envtest.Environment{}
-
-	config, err := env.Start()
-	assert.Nil(t, err)
-
-	defer env.Stop()
-
-	client, err := client.New(config, client.Options{})
-	assert.Nil(t, err)
-
-	// Run the test a bunch of times to try to trigger a conflict and ensure that it handles conflicts properly.
-	for i := 0; i < 10; i++ {
-		depToUpdate := WithSpec(NewPod("update-me", fmt.Sprintf("default-%d", i)), map[string]interface{}{
-			"containers": []map[string]interface{}{
-				{
-					"image": "nginx",
-					"name":  "nginx",
-				},
-			},
-		})
-
-		assert.Nil(t, CreateOrUpdate(context.TODO(), client, SetAnnotation(depToUpdate, "test", "hi"), true))
-
-		quit := make(chan bool)
-
-		go func() {
-			for {
-				select {
-				case <-quit:
-					return
-				default:
-					CreateOrUpdate(context.TODO(), client, SetAnnotation(depToUpdate, "test", fmt.Sprintf("%d", i)), false)
-					time.Sleep(time.Millisecond * 75)
-				}
-			}
-		}()
-
-		time.Sleep(time.Millisecond * 50)
-
-		assert.Nil(t, CreateOrUpdate(context.TODO(), client, SetAnnotation(depToUpdate, "test", "hello"), true))
-
-		quit <- true
 	}
 }
